// frontend/src/api/openapi/models/SentenceAnnotationNodeData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SentenceAnnotationNodeData = {
  /**
   * Background color of the text
   */
  bgcolor: string;
  /**
   * Background color alpha of the text
   */
  bgalpha: number | null;
  type: string;
  /**
   * ID of the sentence annotation
   */
  sentenceAnnotationId: number;
};